spikes/pdf_conversion_spike/requirements.txt
markitdown[pdf]>=0.1.5
pypandoc-binary>=1.17
pdf2docx>=0.5.8
pypdf>=5.4.0
